+++ MEM.Zone-Dashboards/docs/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog for MEM.Zone-Dashboards
 
+## 7.1.0 - 2025-07-30
+
+* Added `Application - Compliance by Collection` report.
+* Added `Package - Compliance by Collection` report.
+* Added `Package - Compliance by Deployment` report.
+* Added `Baseline - Compliance by Collection` report.
+* Added `Software Update - CU Compliance by Collection` report.
+* Fixed and optimized `Clean-SensitveData` build function.
+* Some other minor changes and fixes.
+
 ## 7.0.0 - 2025-06-30
 
 * Fixed `ufn_CM_GetNextMaintenanceWindowForDevice` function `has too many parameters` error [#36](https://github.com/MEM-Zone/MEM.Zone-Dashboards/issues/36). @Zombierbone